requirements.txt
asgiref==3.6.0
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
Django==3.2.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
folium==0.14.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
idna==3.4
importlib-metadata==6.5.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.21.6
PyJWT==2.6.0
pytz==2023.3
ratelim==0.1.6
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0
